refactor: 💡 페이즈, 방에 대한 정의를 common 밑의 definition으로 이동
