+++ backend/src/common/definition/room.ts
@@ -1,0 +1,7 @@
+import { ROOM_PHASE } from './phase';
+
+export type Room = {
+  roomId: string;
+  phase: ROOM_PHASE;
+  hostId: string;
+}
