+++ backend/src/common/guard/phase.guard.ts
@@ -1,0 +1,14 @@
+import { CanActivate, ExecutionContext } from '@nestjs/common';
+import { WsException } from '@nestjs/websockets';
+
+class PhaseGuard implements CanActivate {
+
+  canActivate(context: ExecutionContext): boolean {
+    const client = context.switchToWs().getClient();
+    if (client.rooms.size !== 1) {
+      throw new WsException('이미 접속한 방이 존재합니다.');
+    }
+
+    return true;
+  }
+}
